extract the bulk of the application configuration into a function and separate module so we can push this into the TAAR library
